完善更新状态和运行的代码

+++ StepperStatus.cs
@@ -3,8 +3,8 @@
     internal class StepperStatus
     {
         public int position { get; set; }
-        public int speed { get; set; }
-        public int acceleration { get; set; }
+        public float speed { get; set; }
+        public float acceleration { get; set; }
         public int target { get; set; }
     }
 }
